+++ src/MainPage.css
@@ -2,6 +2,14 @@
   padding-top: 3em;
 }
 
+.yearForm {
+    position: relative;
+    left: 120px;
+    width: 250px;
+    margin-top: 2em;
+    margin-bottom: 1em;
+}
+
 .input-field, .add-form-button {
   padding-bottom: 1em;
 }
